refactor!: rename wrpt to autoport
BREAKING CHANGE: rename this fork from `wrpt` to `autoport`, and
continue development here.
