+++ src/commands/stacks/handlers/list.rs
@@ -2,10 +2,10 @@ use simplelog::debug;
 
 use crate::commands::consts;
 use crate::commands::error::CliError;
+use crate::commands::helpers::CliContext;
 use crate::commands::helpers::build_table;
 use crate::commands::helpers::construct_url;
 use crate::commands::helpers::parse_api_response;
-use crate::commands::helpers::CliContext;
 use crate::commands::stacks::args::list::StackListCommand;
 use crate::commands::stacks::models::list::StackList;
