+++ src/commands/stacks/models/list.rs
@@ -1,13 +1,13 @@
-use chrono::serde::ts_seconds::deserialize as from_ts;
-use chrono::serde::ts_seconds_option::deserialize as from_ts_option;
 use chrono::DateTime;
 use chrono::Utc;
+use chrono::serde::ts_seconds::deserialize as from_ts;
+use chrono::serde::ts_seconds_option::deserialize as from_ts_option;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::commands::stacks::models::resource_control::ResourceControl;
 use crate::commands::stacks::models::StackStatus;
 use crate::commands::stacks::models::StackType;
+use crate::commands::stacks::models::resource_control::ResourceControl;
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "PascalCase")]
